initial outline for new registry resource.  moved old to archive
