+++ archive/ntreg/src/registry_key.rs
