+++ archive/ntreg/src/registry_value.rs
