+++ archive/ntreg/tests/registry_key_tests.rs
