+++ archive/ntreg/tests/registry_value_tests.rs
