+++ archive/registry/src/args.rs
@@ -14,20 +14,11 @@ pub struct Arguments {
 #[derive(Debug, PartialEq, Eq, Subcommand)]
 pub enum ConfigSubCommand {
     #[clap(name = "get", about = "Retrieve registry configuration.")]
-    Get {
-        #[clap(short, long, required = true, help = "The registry JSON input.")]
-        input: String,
-    },
+    Get,
     #[clap(name = "set", about = "Apply registry configuration.")]
-    Set {
-        #[clap(short, long, required = true, help = "The registry JSON input.")]
-        input: String,
-    },
-    #[clap(name = "remove", about = "Delete registry configuration.")]
-    Remove {
-        #[clap(short, long, required = true, help = "The registry JSON input.")]
-        input: String,
-    },
+    Set,
+    #[clap(name = "test", about = "Validate registry configuration.")]
+    Test,
 }
 
 #[derive(Debug, PartialEq, Eq, Subcommand)]
@@ -48,6 +39,8 @@ pub enum SubCommand {
         #[clap(short, long, help = "The value to set.")]
         value: String,
     },
+    #[clap(name = "test", about = "Validate registry matches input JSON.")]
+    Test,
     #[clap(name = "remove", about = "Remove a registry key or value.", arg_required_else_help = true)]
     Remove {
         #[clap(short, long, required = true, help = "The registry key path to remove.")]
@@ -76,5 +69,8 @@ pub enum SubCommand {
         subcommand: ConfigSubCommand,
     },
     #[clap(name = "schema", about = "Retrieve JSON schema.")]
-    Schema,
+    Schema {
+        #[clap(short, long, help = "Pretty print JSON.")]
+        pretty: bool,
+    }
 }
