+++ archive/registry/tests/registry.config.get.tests.ps1
