+++ archive/registry/tests/registry.config.set.tests.ps1
