+++ archive/registry/tests/registry.config.test.tests.ps1
