+++ registry/src/config.rs
@@ -16,10 +16,7 @@ pub enum RegistryValueData {
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, JsonSchema)]
 #[serde(rename = "Registry", deny_unknown_fields)]
-pub struct Registry {
-    /// The ID of the resource.  Value is ignored for input.
-    #[serde(rename = "$id", skip_serializing_if = "Option::is_none")]
-    pub id: Option<String>,
+pub struct RegistryConfig {
     /// The path to the registry key.
     #[serde(rename = "keyPath")]
     pub key_path: String,
@@ -31,44 +28,4 @@ pub struct Registry {
     #[serde(rename = "valueData")]
     #[serde(skip_serializing_if = "Option::is_none")]
     pub value_data: Option<RegistryValueData>,
-    /// Flag indicating whether the registry value should be present or absent.
-    #[serde(rename = "_exist")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub exist: Option<bool>,
-    /// Flag indicating whether the registry value should be overwritten if it already exists.
-    #[serde(rename = "_clobber")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub clobber: Option<bool>,
-    /// Flag indicating whether the resource is in the desired state.  Value is ignored for input.
-    #[serde(rename = "_inDesiredState")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub in_desired_state: Option<bool>,
-}
-
-impl Registry {
-    pub fn to_json(&self) -> String {
-        match serde_json::to_string(self) {
-            Ok(json) => json,
-            Err(e) => {
-                eprintln!("Failed to serialize to JSON: {e}");
-                String::new()
-            }
-        }
-    }
-}
-
-const ID: &str = "https://developer.microsoft.com/json-schemas/windows/registry/20230303/Microsoft.Windows.Registry.schema.json";
-
-impl Default for Registry {
-    fn default() -> Self {
-        Self {
-            id: Some(ID.to_string()),
-            key_path: String::new(),
-            value_name: None,
-            value_data: None,
-            exist: None,
-            clobber: None,
-            in_desired_state: None,
-        }
-    }
 }
